requirements.txt
Flask>=2.0.0
Flask-Cors>=3.0.0
Flask-SQLAlchemy>=2.5.0
mysql-connector-python>=8.0.0
